refactor: Moved feature controls out of Viewer, into component
